pyproject.toml
[tool.poetry]
name = "chatwoot-client"
version = "0.1.1"
description = ""
authors = ["Junior Vidotti <jrvidotti@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.26.0"
pydantic = "^2.6.1"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
